bug fix
this was a secret bug fix. i basically just forgot to remove this piece when i was first building the script. it was causing the script to skip the first property when it shouldn't have been. i don't have to fix the sheet because we've transferred and i fixed the new one.